LayoutTests/media/modern-media-controls/forward-button/forward-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/forward-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ForwardButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const forwardButton = new ForwardButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("forwardButton.element.localName", "button");
shouldBeTrue('forwardButton.element.classList.contains("icon")');
shouldBeTrue('forwardButton.element.classList.contains("forward")');
shouldBeEqualToString("forwardButton.iconName", `${Icons.Forward}`);

scheduler.frameDidFire = function()
{
    if (forwardButton.width == 0)
        return;

    shouldBeTrue('forwardButton.element.style.webkitMaskImage.includes("macOS/forward@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
