LayoutTests/media/modern-media-controls/forward-button/forward-button-expected.txt

Testing the ForwardButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS forwardButton.element.localName is "button"
PASS forwardButton.element.classList.contains("icon") is true
PASS forwardButton.element.classList.contains("forward") is true
PASS forwardButton.iconName is "forward"
PASS forwardButton.element.style.webkitMaskImage.includes("macOS/forward@") became true
PASS successfullyParsed is true

TEST COMPLETE
